+++ assets/ReminderFormAssets.php
@@ -8,15 +8,11 @@
 
 namespace humhub\modules\calendar\assets;
 
-use yii\web\AssetBundle;
+use humhub\components\assets\AssetBundle;
 
 class ReminderFormAssets extends AssetBundle
 {
-    public $defer = true;
-
-    public $publishOptions = [
-        'forceCopy' => false,
-    ];
+    public $forceCopy = false;
 
     public $sourcePath = '@calendar/resources';